pyproject.toml

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "air-quality-monitoring"
version = "0.0.1"
dependencies = [
    "pandas",
    "geopandas",
    "mysql-connector-python",
    "shapely",
    "questdb",
    "pyarrow",
    "requests",
]
requires-python = ">=3.10"
authors = [
    { name = "Daniel Hill", email = "danielmartinhill@gmail.com" },
    { email = "daniel15.hill@live.com" },
]
maintainers = [{ name = "Daniel Hill", email = "danielmartinhill@gmail.com" }]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 100
indent-width = 4

target-version = "py312"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]

quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false
docstring-code-line-length = "dynamic"
